refactor(autoware_grid_map_utils): prefix folder structure with autoware/
