+++ common/autoware_grid_map_utils/include/autoware/grid_map_utils/polygon_iterator.hpp
